+++ src/main/java/HW2/D_Z1.java
@@ -1,4 +1,4 @@
-package Lesson2;
+package HW2;
  // Программа должна считать с консоли число, которое введет пользователь и вывести каждую цифру с новой строки.
 
 import java.util.Scanner;
