+++ src/main/java/HW2/D_Z2.java
@@ -1,4 +1,4 @@
-package Lesson2;
+package HW2;
 //Программа должна подсчитывать и выводить площадь треугольника. Стороны задаются пользователем с консоли.
