+++ src/main/java/HW3/Lesson3_Dz1.java
@@ -1,4 +1,4 @@
-package lesson3;
+package HW3;
 /*C помощью циклов нарисовать такую фигуру:
         *
         **
