+++ src/main/java/HW3/Lesson3_Dz3.java
@@ -1,4 +1,4 @@
-package lesson3;
+package HW3;
     /*Нарисовать обои
         +++***+++***+++***
         +++***+++***+++***
