+++ src/main/java/HW3/Lesson3_Dz5.java
@@ -1,4 +1,4 @@
-package lesson3;
+package HW3;
 
 import java.io.BufferedReader;
 import java.io.InputStreamReader;
