+++ src/main/java/HW3/Lesson3_Dz_4.java
@@ -1,4 +1,4 @@
-package lesson3;
+package HW3;
 
 import java.util.Arrays;
 import java.util.Random;
