+++ src/main/java/HW3/lesson3.java
@@ -1,4 +1,4 @@
-package lesson3;
+package HW3;
 
 import java.lang.reflect.Array;
 import java.util.Arrays;
