+++ src/main/java/HW4/Lesson4_Dz1.java
@@ -1,4 +1,4 @@
-package Lesson4;
+package HW4;
         /*
         ДЗ форматирование:
             Вывести на экран число пи 5 раз, в каждой строке разное количество знак после запятой.
