+++ src/main/java/HW4/Lesson4_Dz2.java
@@ -1,4 +1,4 @@
-package Lesson4;
+package HW4;
 
 import java.util.Arrays;
 import java.util.Scanner;
