+++ src/main/java/HW4/lesson4.java
@@ -1,4 +1,4 @@
-package Lesson4;
+package HW4;
 
 public class lesson4 {
     public static void main(String[] args) {
