+++ src/main/java/HW5/Book.java
@@ -1,4 +1,4 @@
-package Lesson5;
+package HW5;
 /*Создать классы, спецификации которых приведены ниже. Определить конструкторы и методы setТип(), getТип(), toString().
  Определить дополнительно методы в классе, создающем массив объектов. Задать критерий выбора данных и вывести эти данные
   на консоль. В каждом классе, обладающем информацией должно быть объявлено несколько конструкторов.
