+++ src/main/java/HW5/Library.java
@@ -1,4 +1,4 @@
-package Lesson5;
+package HW5;
 import java.util.*;
 import java.util.stream.Collectors;
