+++ src/test/java/HW2/D_Z3Test.java
@@ -1,6 +1,5 @@
-package Lesson2;
+package HW2;
 
-import org.junit.Before;
 import org.junit.Test;
 
 import static org.junit.Assert.*;
